HW-1 : separated Movie from ApiResult
