+++ class13/CmplxN/AAS/app/src/main/java/com/example/aas/ApiResult.kt
@@ -9,15 +9,4 @@ data class ApiResult(
     val lastBuildDate: String,
     val start: Int,
     val total: Int
-) {
-	data class Movie(
-        val actor: String,
-        val director: String,
-        val image: String,
-        val link: String,
-        val pubDate: String,
-        val subtitle: String,
-        val title: String,
-        val userRating: String
-    )
-}
+)
