+++ class13/CmplxN/AAS/app/src/main/java/com/example/aas/Movie.kt
@@ -1,0 +1,12 @@
+package com.example.aas
+
+data class Movie(
+	val actor: String,
+	val director: String,
+	val image: String,
+	val link: String,
+	val pubDate: String,
+	val subtitle: String,
+	val title: String,
+	val userRating: String
+)
